+++ examples/express/src/guestnode.ts
@@ -1,5 +1,5 @@
 import guest from "./guest";
-import { Kademlia, genKid, KvsModule, PeerModule } from "../../..";
+import { Kademlia, genKid, KvsModule, PeerModule } from "../../../src";
 
 const kad = new Kademlia(genKid(), { kvs: KvsModule, peerCreate: PeerModule });
