+++ examples/express/src/portal/index.ts
@@ -1,5 +1,5 @@
 import Express from "express";
-import { Kademlia, PeerModule, Peer } from "../../../..";
+import { Kademlia, PeerModule, Peer } from "../../../../src";
 import bodyParser from "body-parser";
 
 const peers: { [key: string]: Peer } = {};
