+++ examples/express/src/portalnode.ts
@@ -1,5 +1,5 @@
 import portalnode from "./portal";
-import { Kademlia, genKid, KvsModule, PeerModule } from "../../..";
+import { Kademlia, genKid, KvsModule, PeerModule } from "../../../src";
 
 const kad = new Kademlia(genKid(), { kvs: KvsModule, peerCreate: PeerModule });
