+++ examples/socket.io/guest/index.ts
