+++ examples/socket.io/guest/test.ts
