+++ package.json
@@ -25,8 +25,6 @@
     "@types/jest": "^24.0.6",
     "@types/node": "^10.11.7",
     "@types/sha1": "^1.1.1",
-    "@types/socket.io": "^1.4.38",
-    "@types/socket.io-client": "^1.4.32",
     "aport": "^2.0.0",
     "jest": "^24.1.0",
     "ts-jest": "^24.0.0",
@@ -42,8 +40,6 @@
     "keypair": "^1.0.1",
     "rx.mini": "git+https://git@github.com/shinyoshiaki/rx.mini.git",
     "sha1": "^1.1.1",
-    "socket.io": "^2.1.1",
-    "socket.io-client": "^2.1.1",
     "webrtc4me": "^0.5.8",
     "wrtc": "^0.4.0"
   },
